Move `nightly` module to separate file
